Add files via upload
This repository contains the implementation of the methods described in the paper "A Semi-Bayesian Nonparametric Estimator of the Maximum Mean Discrepancy Measure: Applications in Goodness-of-Fit Testing and Generative Adversarial Networks". The code provides tools for estimating the Maximum Mean Discrepancy (MMD) in a semi-Bayesian nonparametric framework, with applications in goodness-of-fit testing and GANs.

The GAN-related code is based on a Bayesian nonparametric (BNP) extension of the method from Li (2015), with modifications adapted from their publicly available code: Generative Moment Matching Networks.